General_Purpose_IoT/raspberry_mqtt_broker/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
influxdb==5.3.2
msgpack==1.1.0
paho-mqtt==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
urllib3==2.4.0
